requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
httpx>=0.27.0
pydantic>=2.9.0
anthropic>=0.18.0
google-genai>=1.64.0
